% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLH.R
\name{filterLH}
\alias{filterLH}
\title{Lyne-Hollick recursive digital filter}
\usage{
filterLH(dataset, alpha = 0.925, passes = 3)
}
\arguments{
\item{dataset}{dataframe - column 1 contains dates, column 2 contains streamflow values}

\item{alpha}{numeric - filter parameter (defaults to 0.925 following Nathan and McMahon (1990))}

\item{passes}{numeric - number of times the filter passes over the data (typically 1-3)}
}
\value{
a dataframe with 3 columns and nrow = length of streamflow dataset. Column one contains streamflow in the same units as input, column two contains baseflow in the same units as input, column three contains quickflow in the same units as input.
}
\description{
Lyne-Hollick recursive digital filter
}
\examples{
filterLH(GreenRiver)
}
