% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUSGS.R
\name{getUSGS}
\alias{getUSGS}
\title{Get USGS Streamgage Discharge Data}
\usage{
getUSGS(gage_no, start_date = "", end_date = "")
}
\arguments{
\item{gage_no}{character - USGS streamgage number, usually an 8-digit number.}

\item{start_date}{character - starting date for data retrieval in the form YYYY-MM-DD. Defaults to earliest possible record.}

\item{end_date}{character - ending date for data retrieval in the form YYYY-MM-DD. Defaults to latest possible record.}
}
\value{
a dataframe with dates and mean daily discharge data (in cfs)
}
\description{
Get USGS Streamgage Discharge Data
}
\examples{
gage_no <- "09315000"
start_date <- "2021-09-30"
end_date <- "2022-10-01"
getUSGS(gage_no, start_date, end_date)
}
