\name{GreenRiver}
\alias{GreenRiver}
\docType{data}
\title{
Discharge Data for the Green River at Green River, UT
}
\description{
A subset of data from USGS streamgage 09315000
}
\usage{data("GreenRiver")}
\format{
  A data frame with 732 observations on the following 2 variables.
  \describe{
    \item{\code{Date}}{a character vector}
    \item{\code{Streamflow_cfs}}{a numeric vector}
  }
}

\source{
<https://waterdata.usgs.gov/monitoring-location/09315000/#parameterCode=00065&period=P7D&showMedian=false>
}

\examples{
head(GreenRiver)
}

\keyword{datasets}
