% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hydrograph.R
\name{hydrograph}
\alias{hydrograph}
\title{Plot a Hydrograph}
\usage{
hydrograph(data)
}
\arguments{
\item{data}{data table -}
}
\value{
a plotted hydrograph
}
\description{
Plot a Hydrograph
}
\examples{

t <- filterLH(GreenRiver)
hydrograph(t)
}
