% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterLH.R
\name{filterLH}
\alias{filterLH}
\title{Title}
\usage{
filterLH(streamflow, alpha = 0.925, passes = 3)
}
\arguments{
\item{streamflow}{vector - streamflow values}

\item{alpha}{numeric - filter parameter (defaults to 0.925 following Nathan and McMahon (1990))}

\item{passes}{numeric - number of times the filter passes over the data (typically 1-3)}
}
\value{
a dataframe with 2 columns and nrow = length of streamflow dataset. Column one contains baseflow in the same units as input, column two contains quickflow in the same units as input.
}
\description{
Title
}
\examples{
streamflow <- GreenRiver[,2]
filterLH(streamflow)
}
